+++ src/default/Min.c
@@ -34,7 +34,7 @@ static void Min_int8(struct onnx_node_t * n)
 	int8_t * py = (int8_t *)y->datas;
 	int8_t * px;
 	int8_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -56,7 +56,7 @@ static void Min_int16(struct onnx_node_t * n)
 	int16_t * py = (int16_t *)y->datas;
 	int16_t * px;
 	int16_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -78,7 +78,7 @@ static void Min_int32(struct onnx_node_t * n)
 	int32_t * py = (int32_t *)y->datas;
 	int32_t * px;
 	int32_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -100,7 +100,7 @@ static void Min_int64(struct onnx_node_t * n)
 	int64_t * py = (int64_t *)y->datas;
 	int64_t * px;
 	int64_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -122,7 +122,7 @@ static void Min_uint8(struct onnx_node_t * n)
 	uint8_t * py = (uint8_t *)y->datas;
 	uint8_t * px;
 	uint8_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -144,7 +144,7 @@ static void Min_uint16(struct onnx_node_t * n)
 	uint16_t * py = (uint16_t *)y->datas;
 	uint16_t * px;
 	uint16_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -166,7 +166,7 @@ static void Min_uint32(struct onnx_node_t * n)
 	uint32_t * py = (uint32_t *)y->datas;
 	uint32_t * px;
 	uint32_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -188,7 +188,7 @@ static void Min_uint64(struct onnx_node_t * n)
 	uint64_t * py = (uint64_t *)y->datas;
 	uint64_t * px;
 	uint64_t minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -211,7 +211,7 @@ static void Min_bfloat16(struct onnx_node_t * n)
 	uint16_t * px;
 	float v;
 	float minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -235,7 +235,7 @@ static void Min_float16(struct onnx_node_t * n)
 	uint16_t * px;
 	float v;
 	float minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -258,7 +258,7 @@ static void Min_float32(struct onnx_node_t * n)
 	float * py = (float *)y->datas;
 	float * px;
 	float minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
@@ -280,7 +280,7 @@ static void Min_float64(struct onnx_node_t * n)
 	double * py = (double *)y->datas;
 	double * px;
 	double minv;
-	int i, j, l;
+	size_t i, j, l;
 
 	for(i = 0, l = y->ndata; i < l; i++)
 	{
